+++ packages/components/rollup.config.mjs
@@ -6,7 +6,7 @@ export default rollupConfig({
     future: './src/__rc__/index.ts',
     actionsV1: './src/v1-actions.ts',
     actionsV2: './src/v2-actions.ts',
-    actionsV3: './src/__actions__/v3.ts',
+    actionsV3: './src/v3-actions.ts',
     overlaysV1: './src/v1-overlays.ts',
     overlaysV2: './src/v2-overlays.ts',
     overlaysV3: './src/v3-overlays.ts',
