+++ packages/components/src/__actions__/Menu/v3/_docs/Menu.docs.stories.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent, ReactNode } from 'react'
 import { Meta, StoryObj } from '@storybook/react'
 import isChromatic from 'chromatic'
-import { Button } from '~components/__actions__/v3'
+import { Button } from '~components/__rc__/Button'
 import { Icon } from '~components/__rc__/Icon'
 import { Menu, MenuItem, MenuTrigger, MenuPopover, MenuSection, MenuHeader } from '../index'
