+++ packages/components/src/__actions__/Menu/v3/_docs/Menu.stories.tsx
@@ -2,7 +2,7 @@ import React, { FunctionComponent } from 'react'
 import { Meta, StoryObj } from '@storybook/react'
 import isChromatic from 'chromatic'
 import { Text } from '~components/Text'
-import { Button } from '~components/__actions__/v3'
+import { Button } from '~components/__rc__/Button'
 import { Icon } from '~components/__rc__/Icon'
 import { Menu, MenuTrigger, MenuItem, MenuPopover, MenuSection, MenuHeader } from '../index'
 import * as testStories from './Menu.spec.stories'
