+++ packages/components/src/__actions__/v3.ts
@@ -1,2 +1,1 @@
 export * from './Menu/v3'
-export * from './Button/v3'
