+++ packages/components/src/__rc__/Button/_docs/Button--api-specification.mdx
@@ -10,7 +10,7 @@ import * as specStories from './Button.spec.stories'
 Updated Nov 19, 2024
 
 <ResourceLinks
-  sourceCode="https://github.com/cultureamp/kaizen-design-system/tree/main/packages/components/src/__actions__/Button/v3"
+  sourceCode="https://github.com/cultureamp/kaizen-design-system/tree/main/packages/components/src/__rc__/Button"
   figma="https://www.figma.com/design/eZKEE5kXbEMY3lx84oz8iN/branch/sPhYSlgPScLOAOYfAbkaI5/%F0%9F%92%9C-Heart-UI-Kit?node-id=1929-17364&node-type=canvas&m=dev"
   designGuidelines="/?path=/docs/actions-button-button-v3-usage-guidelines--docs"
 />
