+++ packages/components/src/__rc__/Button/subcomponents/PendingContent/PendingContent.module.css
