+++ packages/components/src/__rc__/Button/subcomponents/PendingContent/index.ts
