+++ packages/components/src/__rc__/Icon/_docs/Icon.stickersheet.stories.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { Meta } from '@storybook/react'
-import { Button as ButtonV2 } from '~components/Button'
+import { Button } from '~components/Button'
 import { CheckboxField } from '~components/Checkbox'
 import { Label } from '~components/Label'
 import { Text } from '~components/Text'
-import { Button as ButtonV3 } from '~components/__actions__/v3'
+import { Button as ButtonRC } from '~components/__rc__/Button'
 import { StickerSheet, StickerSheetStory } from '~storybook/components/StickerSheet'
 import { handledRtlIcons, iconDefaultSet } from '../constants'
 import { Icon, IconProps } from '../index'
@@ -87,12 +87,12 @@ const StickerSheetTemplate: StickerSheetStory = {
             />
           </StickerSheet.Row>
           <StickerSheet.Row header="Button (v2)">
-            <ButtonV2 label="Button label" icon={<Icon {...defaultProps} />} />
+            <Button label="Button label" icon={<Icon {...defaultProps} />} />
           </StickerSheet.Row>
           <StickerSheet.Row header="Button (v3)">
-            <ButtonV3>
+            <ButtonRC>
               <Icon {...defaultProps} /> Button label
-            </ButtonV3>
+            </ButtonRC>
           </StickerSheet.Row>
         </StickerSheet>
       </>
