+++ packages/components/src/__rc__/index.ts
@@ -1,3 +1,4 @@
+export * from './Button'
 export * from './Icon'
 export * from './Select'
 export * from './Tag'
