+++ packages/components/src/v3-actions.ts
@@ -1,0 +1,1 @@
+export * from './__rc__/Button'
